% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{dynamic_knapsack}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{A DataFrame}

\item{W}{an Integer}
}
\value{
A list
}
\description{
dynamic_knapsack
}
\examples{
dynamic_knapsack(x = knapsack_objects[1:8,], W = 3500)
dynamic_knapsack(x = knapsack_objects[1:12,], W = 3500)
dynamic_knapsack(x = knapsack_objects[1:8,], W = 2000)
dynamic_knapsack(x = knapsack_objects[1:12,], W = 2000)
}
