% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knapsack_objects}
\alias{knapsack_objects}
\title{Knapsack objects
A dataframe use as ain input to solve napsack problem}
\format{
\describe{
\item{v}{Numeric value }
\item{w}{Numeric weight }}
}
\usage{
knapsack_objects
}
\description{
Knapsack objects
A dataframe use as ain input to solve napsack problem
}
\keyword{datasets}
